Moved [MaybeNullTaskResult] to NullGuard.CodeAnalysis namespace
Otherwise collisions will happen between NullGuard.AllowNull and System.Diagnostics.CodeAnalysis.AllowNullAttribute
