+++ AssemblyToProcessNullableRefTypes/ClassWithAsyncMethods.cs
@@ -1,10 +1,8 @@
 ﻿#pragma warning disable CS1998 // Async method lacks 'await' operators and will run synchronously
 
 using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
-using NullGuard;
+using NullGuard.CodeAnalysis;
 
 public class ClassWithAsyncMethods
 {
